Adjust for three-digit CHAIN_PORT prefixes
Most ports are being moved from `${CHAIN_PORT}123` to `${CHAIN_PORT}23`
where the last two digits are retained. The Prefix value itself is the
previously used two-digit value with an additional `1` added at start
so that we do not need to venture into the 40k+ port range, which
already starts to overlap with randomly allocated ports.

Considering that P2P port `060` becomes `60`, we are moving the
Prometheus port (previously `660`) to a different value: `66` instead.